CC-25965:renamed files
