--- perfomance-tests/environments/b2c.json
@@ -1,8 +1,0 @@
-{
-    "local": {
-        "baseUrl": "http://host.docker.internal:9191"
-    },
-    "testing": {
-        "baseUrl": "https://www.de.mp-b2b.internal-testing.demo-spryker.com"
-    }
-}
