+++ perfomance-tests/helpers/storefront-helper.js
