+++ perfomance-tests/tests/b2b-mp/abstract-b2b-mp-scenario.js
@@ -1,4 +1,4 @@
-import { AbstractScenario } from '../AbstractScenario.js';
+import { AbstractScenario } from '../abstract-scenario.js';
 
 export class AbstractB2bMpScenario extends AbstractScenario {
     ENVIRONMENT() {
